+++ app/src/main/java/com/example/weather/DetailsFragment.kt
@@ -9,13 +9,12 @@ import androidx.annotation.RequiresApi
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import com.bumptech.glide.Glide
+import com.example.weather.CityPickerViewModel.Companion.convertCelsiusToFahrenheit
+import com.example.weather.DetailsViewModel.Companion.convertTime
+import com.example.weather.DetailsViewModel.Companion.formatLowAndHighDetails
 import com.example.weather.databinding.FragmentDetailsBinding
 import com.example.weather.network.CityDetails
 import com.example.weather.network.StateDetails
-import java.time.Instant
-import java.time.ZoneOffset
-import java.time.format.DateTimeFormatter
-import java.util.Locale
 
 internal const val base_url = "https://openweathermap.org/img/wn/"
 internal const val url_suffix = "@2x.png"
@@ -70,27 +69,6 @@ class DetailsFragment : Fragment() {
     override fun onDestroyView() {
         super.onDestroyView()
         _binding = null
-    }
-
-
-    private fun formatLowAndHighDetails(it: CityDetails?): String {
-        val low = convertCelsiusToFahrenheit(it?.main?.temp_min).toString()
-        val high = convertCelsiusToFahrenheit(it?.main?.temp_max).toString()
-        return "$low\u2109/$high\u2109"
-    }
-
-    private fun convertCelsiusToFahrenheit(temp: Double?): Int {
-        return if (temp != null) {
-            return (temp * 9/5 + 32).toInt()
-        } else 0
-    }
-
-    @RequiresApi(Build.VERSION_CODES.O)
-    fun convertTime(timestamp: Int?, timeZoneOffset: Int?): String {
-        val offset = ZoneOffset.ofTotalSeconds(timeZoneOffset!!)
-        val instant = Instant.ofEpochSecond(timestamp!!.toLong())
-        val formatter = DateTimeFormatter.ofPattern("K:mm a", Locale.ENGLISH)
-        return instant.atOffset(offset).format(formatter)
     }
